+++ backend/src/constants/routes.constants.ts
@@ -4,3 +4,17 @@ export const TASK_ROUTES = {
   UPDATE_TASK: `/tasks/:id/`,
   DELETE_TASK: `/tasks/:id/`,
 };
+
+export const POST_ROUTES = {
+  GET_MY_POSTS: `/posts`,
+  CREATE_POST: `/posts`,
+  UPDATE_POST: `/posts/:id/`,
+  DELETE_POST: `/posts/:id/`,
+};
+
+export const TAG_ROUTES = {
+  GET_MY_TAGS: `/tags`,
+  CREATE_TAG: `/tags`,
+  UPDATE_TAG: `/tags/:id/`,
+  DELETE_TAG: `/tags/:id/`,
+};