+++ backend/src/constants/strings.constants.ts
@@ -1,0 +1,6 @@
+export const ERRORS = {
+  unknownError: "An unknown error occurred",
+  posts: {
+    notFound: "The post not found",
+  },
+};