+++ backend/src/loaders/express.loader.ts
@@ -1,9 +1,12 @@
-import express, { Express, Request, Response } from "express";
+import express, { Express } from "express";
+import appRoutes from "../routes";
 
 export default function loadServer() {
   const app: Express = express();
 
   app.use(express.json());
+
+  appRoutes(app);
 
   return app;
 }