+++ backend/src/models/posts/post.model.ts
@@ -6,7 +6,7 @@ import { PaginatedModel } from "../types";
 const PostSchema = new mongoose.Schema<Post>(
   {
     title: { type: String, required: true },
-    post: { type: String },
+    content: { type: String },
     isArchived: { type: Boolean, default: false },
     tags: [{ type: mongoose.Schema.Types.ObjectId, ref: "Tag" }],
   },