+++ backend/src/models/posts/types.d.ts
@@ -3,7 +3,7 @@ import { Tag } from "../tags/types";
 
 export interface Post {
   title: string;
-  post: string;
+  content: string;
   isArchived: boolean;
   tags: Tag[];
 }